chore: moved from general exceptions to apis/common/exceptions.py
